auto-test-framework: add UpdatePetTest;

+++ auto-test-framework/src/main/java/io/swagger/petstore/model/builder/PetBuilder.java
@@ -14,7 +14,7 @@ public class PetBuilder {
         pet = new Pet();
     }
 
-    public PetBuilder setId(int id) {
+    public PetBuilder setId(long id) {
         pet.setId(id);
         return this;
     }
